+++ RandomMoonFX/Plugin.cs
@@ -8,24 +8,33 @@ namespace RandomMoonFX
 {
     [BepInPlugin(GUID, NAME, VERSION)]
     [BepInDependency("com.github.darmuh.LethalConstellations", BepInDependency.DependencyFlags.SoftDependency)]
+    [BepInDependency("butterystancakes.lethalcompany.chameleon", BepInDependency.DependencyFlags.SoftDependency)]
+    [BepInDependency("CelestialTint", BepInDependency.DependencyFlags.SoftDependency)]
+    [BepInDependency("JacobG5.WesleyMoons", BepInDependency.DependencyFlags.SoftDependency)]
+    [BepInDependency("CodeRebirth", BepInDependency.DependencyFlags.SoftDependency)]
     public class Plugin : BaseUnityPlugin
     {
         const string GUID = "zigzag.randommoonfx";
         const string NAME = "RandomMoonFX";
-        const string VERSION = "1.3.5";
+        const string VERSION = "1.3.6";
 
         public static Plugin instance;
         private readonly Harmony harmony = new Harmony(GUID);
         internal static Config config { get; private set; } = null!;
 
         private readonly int GordionID = 3;
-        public int GaletryID = -99;
+        public int ModdedCompanyID = -99;
+        public bool IsGaletryCompany = false;
+        public bool IsOxydeCompany = false;
+        public int NbOfCompanyMoons = 1;
         public float AnimationTime = 1.5f;
         public bool IsStarting = false;
         public Terminal? terminal;
         public int terminalCostOfItems = -5;
         public bool constellationsCompatibility = false;
         public List<string> VisitedMoons = new List<string>();
+
+        public string ActualCompanyName => ModdedCompanyID == -99 ? "71 Gordion" : IsOxydeCompany ? "745-Oxyde" : "98 Galetry";
 
         [System.Diagnostics.CodeAnalysis.SuppressMessage("CodeQuality", "IDE0051")]
         void Awake()
@@ -54,8 +63,25 @@ namespace RandomMoonFX
                 AnimationTime = 4f;
             if (config.AnimationTimeOverride.Value >= 0f)
                 AnimationTime = config.AnimationTimeOverride.Value;
-            if (config.GaletryCompany.Value && Chainloader.PluginInfos.ContainsKey("JacobG5.WesleyMoons"))
-                GaletryID = -1;
+            if (Chainloader.PluginInfos.ContainsKey("JacobG5.WesleyMoons"))
+            {
+                NbOfCompanyMoons++;
+                if (config.GaletryCompany.Value)
+                {
+                    ModdedCompanyID = -1;
+                    IsGaletryCompany = true;
+                }
+            }
+            if (Chainloader.PluginInfos.ContainsKey("CodeRebirth"))
+            {
+                NbOfCompanyMoons++;
+                if (config.OxydeCompany.Value)
+                {
+                    ModdedCompanyID = -1;
+                    IsOxydeCompany = true;
+                    IsGaletryCompany = false;
+                }
+            }
             if (config.ConstellationsCheck.Value && Chainloader.PluginInfos.ContainsKey("com.github.darmuh.LethalConstellations"))
                 constellationsCompatibility = true;
         }
@@ -74,27 +100,27 @@ namespace RandomMoonFX
         {
             if (LastDayOfQuota())
             {
-                if (GaletryID != -99)
+                if (ModdedCompanyID != -99)
                 {
-                    if (GaletryID == -1)
+                    if (ModdedCompanyID == -1)
                     {
                         foreach (var level in StartOfRound.Instance.levels)
                         {
-                            if (level.PlanetName == "98 Galetry")
+                            if (level.PlanetName == ActualCompanyName)
                             {
-                                GaletryID = level.levelID;
+                                ModdedCompanyID = level.levelID;
                                 break;
                             }
                         }
                     }
-                    if (GaletryID != -1)
+                    if (ModdedCompanyID != -1)
                     {
-                        Logger.LogInfo("Force navigating to 98 Galetry");
-                        StartOfRound.Instance.ChangeLevelServerRpc(GaletryID, FindObjectOfType<Terminal>().groupCredits);
+                        Logger.LogInfo("Force navigating to " + ActualCompanyName);
+                        StartOfRound.Instance.ChangeLevelServerRpc(ModdedCompanyID, FindObjectOfType<Terminal>().groupCredits);
                         IsStarting = true;
                         return;
                     }
-                    Logger.LogError("Galetry Moon was not found");
+                    Logger.LogError("Modded Company Moon was not found");
                 }
                 Logger.LogInfo("Force navigating to the Company Building");
                 StartOfRound.Instance.ChangeLevelServerRpc(GordionID, FindObjectOfType<Terminal>().groupCredits);